Remove bad test of option overriding, and replace with a working one (fixes RT#158253)
The old test never did the right thing -- it is trying to test functionality
that never existed and was never documented to exist. However, this has only just
now been discovered because:

- Getopt::Long::Descriptive always enables no_ignore_case
- there were no tests for turning no_ignore_case off (and indeed, if one
  writes them, one would get the error "Duplicate specification "SmallD|d!"
  for option "d""), because the "small d" and "big D" attributes would map to
  the same command line option
- Getopt::Long::Descriptive fails* to restore state after changing
  configurations as part of describe_options()
- and the GLD tests were being run ahead of the Basic ones

This resulted in the tests conveniently passing, even though the supposed
functionality did not exist, because...

*Getopt::Long::Descriptive (as of version 0.116) has fixed the global state
issue, by capturing the original set of Getopt::Long configs before generating
the option spec, and then restoring them afterwards.

The proper way to change Getopt::Long configs is via the getopt_conf parameter
passed to the MooseX::Getopt::GLD parameterized role, so a new test has been
added to explicitly cover that functionality.